+++ packages/side-toolbar/src/theme.ts
@@ -5,8 +5,9 @@ export interface SideToolbarPluginTheme {
   buttonStyles?: DraftJsButtonTheme;
   blockTypeSelectStyles?: {
     blockType?: string;
-    spacer?: string;
     popup?: string;
+    arrow?: string;
+    popupFrame?: string;
   };
   toolbarStyles?: {
     wrapper?: string;
@@ -63,50 +64,77 @@ const blockTypeSelectStyles: SideToolbarPluginTheme['blockTypeSelectStyles'] = {
       fill: #888;
     }
   `,
-
-  spacer: css`
-    position: absolute;
-    left: 50%;
-    transform: translate(-50%);
-    width: 74px;
-    height: 8px;
-  `,
-
-  popup: css`
-    position: absolute;
-    left: 50%;
-    transform: translate(-50%);
-    background: #efefef;
+  popupFrame: css`
     border: 1px solid #ddd;
     background: #fff;
     border-radius: 2px;
     box-shadow: 0px 1px 3px 0px rgba(220, 220, 220, 1);
-    z-index: 3;
     box-sizing: border-box;
     width: 74px;
-    margin-top: 8px;
-    &:after,
-    &:before {
-      bottom: 100%;
-      left: 50%;
-      border: solid transparent;
-      content: ' ';
-      height: 0;
-      width: 0;
+  `,
+  popup: css`
+    z-index: 3;
+    padding: 8px;
+    visibility: hidden;
+    opacity: 0;
+    transition: opacity 0.25s cubic-bezier(0.3, 1.2, 0.2, 1);
+    &:hover,
+    &:hover > div::before,
+    &[data-show='true'] {
+      visibility: visible;
+      opacity: 1;
+    }
+  `,
+
+  arrow: css`
+    &,
+    &::before {
       position: absolute;
-      pointer-events: none;
+      width: 6px;
+      height: 6px;
+      background: inherit;
     }
-    &:after {
-      border-color: rgba(251, 251, 251, 0);
-      border-bottom-color: #fbfbfb;
-      border-width: 4px;
-      margin-left: -4px;
+
+    &::before {
+      content: '';
+      border: 1px solid #ddd;
+      transform: rotate(45deg);
     }
-    &:before {
-      border-color: rgba(221, 221, 221, 0);
-      border-bottom-color: #ddd;
-      border-width: 6px;
-      margin-left: -6px;
+
+    &[data-popper-placement^='top'] {
+      bottom: 4px;
+    }
+
+    &[data-popper-placement^='bottom'] {
+      top: 4px;
+    }
+
+    &[data-popper-placement^='left'] {
+      right: 4px;
+    }
+
+    &[data-popper-placement^='right'] {
+      left: 4px;
+    }
+
+    &[data-popper-placement^='top']::before {
+      border-left: 0;
+      border-top: 0;
+    }
+
+    &[data-popper-placement^='bottom']::before {
+      border-right: 0;
+      border-bottom: 0;
+    }
+
+    &[data-popper-placement^='left']::before {
+      border-left: 0;
+      border-bottom: 0;
+    }
+
+    &[data-popper-placement^='right']::before {
+      border-right: 0;
+      border-top: 0;
     }
   `,
 };